Replace monolithic e2e-ocl test suite with sharded suite
Replace the monolithic e2e-ocl test suite with two sharded suites
(e2e-ocl-1of2 and e2e-ocl-2of2) and a shared helper package
(e2e-ocl-shared) to reduce overall CI time by running in parallel.

The sharded tests incorporate all stabilization fixes from main:
- Fix cleanup verification timeouts with per-resource contexts
- Fix cleanup ordering to delete owners before dependents
- Increase poll intervals to avoid rate limiter exhaustion
- Make log streaming errors non-fatal
- Increase job completion timeout for CI environments
- Make TestImagePrunerErrors flexible to registry behavior changes

Adapt MachineOSBuild construction calls to use the
NewMachineOSBuildFromAPIOrDie API available on this branch.

Exclude TestStaleAnnotationClearedOnLayerOnlyChange as it requires
a corresponding production fix not present on this branch.

Makefile targets:
- Remove: test-e2e-ocl (190m)
- Add: test-e2e-ocl-1of2 (120m), test-e2e-ocl-2of2 (150m)
