+++ test/e2e-ocl-1of2/main_test.go
@@ -1,0 +1,16 @@
+package e2e_ocl_1of2_test
+
+import (
+	"flag"
+	"os"
+	"testing"
+
+	"k8s.io/klog/v2"
+)
+
+func TestMain(m *testing.M) {
+	flag.Parse()
+	klog.Infof("-skip-cleanup: %v", skipCleanupAlways)
+	klog.Infof("-skip-cleanup-on-failure: %v", skipCleanupOnlyAfterFailure)
+	os.Exit(m.Run())
+}
