+++ test/e2e-ocl-shared/containerfiles.go
@@ -1,0 +1,30 @@
+package e2e_ocl_shared_test
+
+import (
+	_ "embed"
+)
+
+var (
+	// Provides a Containerfile that installs cowsayusing the Centos Stream 9
+	// EPEL repository to do so without requiring any entitlements.
+	//go:embed Containerfile.cowsay
+	CowsayDockerfile string
+
+	// Provides a Containerfile that installs Buildah from the default RHCOS RPM
+	// repositories. If the installation succeeds, the entitlement certificate is
+	// working.
+	//go:embed Containerfile.entitled
+	EntitledDockerfile string
+
+	// Provides a Containerfile that works similarly to the cowsay Dockerfile
+	// with the exception that the /etc/yum.repos.d and /etc/pki/rpm-gpg key
+	// content is mounted into the build context by the BuildController.
+	//go:embed Containerfile.yum-repos-d
+	YumReposDockerfile string
+
+	//go:embed Containerfile.okd-fcos
+	OkdFcosDockerfile string
+
+	//go:embed Containerfile.simple
+	SimpleDockerfile string
+)
